+++ Source/FunicularSwitch.Generators/Transformer/ConstructType.cs
@@ -1,0 +1,3 @@
+namespace FunicularSwitch.Generators.Transformer;
+
+internal delegate TypeInfo ConstructType(IReadOnlyList<TypeInfo> typeParameters);
